% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprod.r
\name{tcrossprod_sparse}
\alias{tcrossprod_sparse}
\title{The tcrossprod function for sparse matrices, with output filters applied on the fly to reduce memory usage.}
\usage{
tcrossprod_sparse(m1, m2 = NULL, min_value = 0, only_upper = F,
  diag = T, top_n = NULL, verbose = F)
}
\arguments{
\item{m1}{A dgCMatrix}

\item{m2}{A dgCMatrix}

\item{min_value}{a numerical value, specifying the threshold for including a score in the output}

\item{only_upper}{if true, only the upper triangle of the matrix is returned. Only possible for symmetrical output (m1 and m2 have same number of columns)}

\item{diag}{if false, the diagonal of the matrix is not returned. Only possible for symmetrical output (m1 and m2 have same number of columns)}

\item{top_n}{an integer, specifying the top number of strongest scores for each column in m1}

\item{verbose}{if TRUE, report progress}
}
\value{
A dgCMatrix
}
\description{
The tcrossprod function for sparse matrices, with output filters applied on the fly to reduce memory usage.
}
\examples{
set.seed(1)
m = Matrix::rsparsematrix(5,10,0.5)
tcrossprod_sparse(m, min_value = 0, only_upper = F, diag = T)
tcrossprod_sparse(m, min_value = 0, only_upper = F, diag = F)
tcrossprod_sparse(m, min_value = 0, only_upper = T, diag = F)
tcrossprod_sparse(m, min_value = 0.2, only_upper = T, diag = F)
tcrossprod_sparse(m, min_value = 0, only_upper = T, diag = F, top_n = 1)
}
